requirements.txt
Flask
gspread
google-auth
google-api-python-client
pandas
matplotlib